🎨 improve screen time performances
